requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
PyMuPDF==1.18.17
Pillow==8.3.2